Refactor loot generation event
